move db seeder to configuration package, make classes package private
