+++ codebase/api/api-person/src/main/java/com/jarqprog/personapi/configuration/SeedingPersonDomain.java
@@ -1,0 +1,7 @@
+package com.jarqprog.personapi.configuration;
+
+interface SeedingPersonDomain {
+
+    void populateDb();
+
+}
