--- codebase/api/api-person/src/main/java/com/jarqprog/personapi/seederdb/SeedingPersonDomain.java
@@ -1,7 +1,0 @@
-package com.jarqprog.personapi.seederdb;
-
-public interface SeedingPersonDomain {
-
-    void populateDb();
-
-}
